More features for geometry definition
FreeCad files created with previous version of Beltrami are not compatible with this version.
Warning messages on non monotonic curves have been clarified. Now, the curves with problems are identified.
The thickness distribution law is now more flexible thanks to the addition of nodes in sketchs (skLoiEpaisseur1e and followings) and variables in Tableau_pilote.